pyproject.toml
[project]
name = "stock-bot"
version = "0.1.0"
description = "Automated stock universe fetching and clustering analysis for Chinese exchanges"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "stock-bot team" }]
keywords = ["stock", "finance", "clustering", "analysis"]

dependencies = [
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "pydantic>=2.5.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "orjson>=3.9.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[project.scripts]
stock-bot = "src.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.sdist]
include = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
